last

+++ app.py
@@ -7,7 +7,8 @@ import threading
 import cherrypy
 
 rootpath  = os.path.abspath(os.path.dirname(__file__))
-
+#this is the default, it will try to get env var defined in apache conf file see chpy.conf.apache-example
+appcontext = "chpy"
 
 cherrypy.config.update({'environment': 'embedded'})
 
@@ -17,9 +18,7 @@ if cherrypy.__version__.startswith('3.0') and cherrypy.engine.state == 0:
 
 def authUser(user, passw):
 	import ldap
-	SERVER_NAME = "ferrari.cca.iac.es"
-	BASE_USER = "ou=People,dc=cca,dc=iac,dc=es"
-	USER_ATTR = "uid"
+	from ldap_config import SERVER_NAME, BASE_USER, USER_ATTR
 	try:
 		l = ldap.open(SERVER_NAME)
 		l.protocol_version = ldap.VERSION3	
@@ -47,7 +46,7 @@ class Root(object):
 		if(authUser(username, password)):
 			cherrypy.session['username'] = username
 			cherrypy.log("HAS USERNAME KEY %s" % str(cherrypy.session.get("username")))
-			raise cherrypy.HTTPRedirect("/chpy")
+			raise cherrypy.HTTPRedirect("/%s" % appcontext)
 		else:
 			return open(os.path.join(rootpath,"html", "login.html"))
 	
@@ -85,15 +84,16 @@ class Root(object):
 			pmfilename = Template(FORWARD_PROCMAIL_FILE).safe_substitute(username = username)	
 			msgfilename = Template(FORWARD_MSG_FILE).safe_substitute(username = username)	
 			ffile = open(ffilename, "w")
-			ffile.write(env.get_template('forward').render(username=username))
+			#TODO replace program binary path in templates
+			ffile.write(env.get_template('forward').render(username=username, procfilepath=pmfilename))
 			ffile.close()
 			pmfile = open(pmfilename, "w")
-			pmfile.write(env.get_template('procmail').render(username=username))
+			pmfile.write(env.get_template('procmail').render(username=username, msgfilepath=msgfilename))
 			pmfile.close()
 			msgfile= open(msgfilename, "w")
 			msgfile.write(msg)
 			msgfile.close()
-		raise cherrypy.HTTPRedirect("/chpy")
+		raise cherrypy.HTTPRedirect("/%s" % appcontext)
 
 	@cherrypy.expose
 	def delete(self,action):
@@ -104,23 +104,25 @@ class Root(object):
 			ffilename = Template(FORWARD_FILE).safe_substitute(username = username)	
 			pmfilename = Template(FORWARD_PROCMAIL_FILE).safe_substitute(username = username)	
 			msgfilename = Template(FORWARD_MSG_FILE).safe_substitute(username = username)	
-			os.remove(ffilename)
-			os.remove(pmfilename)
-			if(action=="del"):
+			if(os.path.isfile(ffilename)):
+				os.remove(ffilename)
+			if(os.path.isfile(pmfilename)):
+				os.remove(pmfilename)
+			if(action=="del" and os.path.isfile(msgfilename)):
 				os.remove(msgfilename)
-		raise cherrypy.HTTPRedirect("/chpy")
+		raise cherrypy.HTTPRedirect("/%s" % appcontext)
 
 
 	@cherrypy.expose
 	def logout(self):
 		cherrypy.session.clear()
-		raise cherrypy.HTTPRedirect("/chpy")
+		raise cherrypy.HTTPRedirect("/%s" % appcontext)
 
 routesConf = {
 		'/': {
 		    'tools.sessions.on': True,
 			  'tools.sessions.storage_type': "file",
-    	  'tools.sessions.storage_path': "/var/www/chpy/sessions",
+    	  'tools.sessions.storage_path': os.path.join(rootpath,"sessions"),
         'tools.sessions.timeout': 600 
 		    #'tools.staticdir.root': os.path.abspath(os.getcwd())
 		},
@@ -134,9 +136,13 @@ routesConf = {
 def application(environ, start_response):
 		cherrypy.config.update({
 			#'environment': 'production',
-		                    'log.error_file': '/var/www/chpy/log/site.log',
+		                    'log.error_file': os.path.join(rootpath, 'log', 'site.log'),
 		                    # ...
 		                    })
-		cherrypy.tree.mount(Root(), script_name="/chpy", config=routesConf)
+		if 'appcontext' in environ:
+			appcontext = environ["appcontext"]
+			cherrypy.log("got env var appcontext")
+		cherrypy.log("APPCONTEXT has value %s" % appcontext)
+		cherrypy.tree.mount(Root(), script_name="/%s" % appcontext, config=routesConf)
 		return cherrypy.tree(environ, start_response)
